cambios botones

+++ app/Http/Controllers/CobrosController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Cobro;
 use App\Models\Comanda;
+use App\Models\Zona;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class CobrosController extends Controller
@@ -13,31 +15,43 @@ class CobrosController extends Controller
      */
     public function index()
     {
-        
+        if (Auth::check()) {
+            $cobros = Cobro::all();
+            $zonas = Zona::all();
+            $total = 0;
+            foreach ($cobros as $cobro) {
+
+                $total += $cobro->cantidad;
+            }
+
+            return view('cobros.index', compact('cobros', 'total', 'zonas'));
+        }
+        return view('welcome');
     }
 
     /**
      * Store a newly created resource in storage.
      */
 
-   
-        public function store(Request $request)
-        {
 
-                Cobro::create($request->all());
+    public function store(Request $request)
+    {
+        if (Auth::check()) {
+            Cobro::create($request->all());
 
-                $comandas = Comanda::where('mesa', $request->mesa)
+            $comandas = Comanda::where('mesa', $request->mesa)
                 ->where('zona_id', $request->zona_id)->get();
-    
-    
-       
-                foreach ($comandas as $comanda){
-               
-                    $comanda->delete();
-                }
 
-                return view('home');
-          
+
+
+            foreach ($comandas as $comanda) {
+
+                $comanda->delete();
+            }
+
+            return view('home');
+        }
+        return view('welcome');
     }
